pyproject.toml
[project]
name = "face-id"
version = "1.0.0"
description = "Système de reconnaissance faciale avec InsightFace et FAISS"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "insightface>=0.7.3",
    "onnxruntime-gpu>=1.18.0; platform_system != 'Darwin'",
    "onnxruntime>=1.18.0; platform_system == 'Darwin'",
    "faiss-cpu>=1.8.0",
    "opencv-python>=4.9.0.80",
    "numpy>=1.26",
    "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.flake8]
max-line-length = 100
exclude = [".git", "__pycache__", ".venv"]
